refactor: refactored 'Color-Scheme-Switcher' logic
