+++ src/js/modules/color-scheme-switcher/config/index.ts
@@ -1,0 +1,4 @@
+export interface ISetActiveControlProps {
+  controls: HTMLElement[]
+  colorScheme: string
+}
