+++ src/js/modules/color-scheme-switcher/index.ts
@@ -1,0 +1,3 @@
+import ColorSchemeSwitcher from './ui/ColorSchemeSwitcher'
+
+export default ColorSchemeSwitcher
