+++ src/js/modules/color-scheme-switcher/ui/ColorSchemeSwitcher.ts
@@ -1,14 +1,18 @@
-import type { NullishString, UndefinedishHTMLElem } from '../types'
+import * as _ from '../config'
+import type { NullishString, UndefinedishHTMLElem } from '../../../types'
 
-interface ISetActiveControlProps {
-  controls: HTMLElement[]
-  colorScheme: string
-}
-
+/**
+ * @class
+ * @module ColorSchemeSwitcher
+*/
 export default class ColorSchemeSwitcher {
   private _switcher: HTMLElement
   private _controls: NodeListOf<HTMLElement>
 
+  /**
+   * @constructor
+   * @param {HTMLElement} colorSchemeSwitcher
+  */
   constructor(colorSchemeSwitcher: HTMLElement) {
     this._switcher = colorSchemeSwitcher
 
@@ -19,6 +23,10 @@ export default class ColorSchemeSwitcher {
     this._controls = this._switcher.querySelectorAll('[data-color-scheme-switcher="control"]')
   }
 
+  /**
+   * @description Public method init(). Used to initiate the 'Color-Scheme-Switcher' logic.
+   * @returns {void}
+  */
   public init(): void {
     const initialColorScheme: string = this._getColorSchemeFromLocalStorage() ?? 'dark'
     this._setColorScheme(initialColorScheme)
@@ -28,15 +36,30 @@ export default class ColorSchemeSwitcher {
     })
   }
 
+  /**
+   * Private method: [_getColorSchemeFromLocalStorage()]
+   * @returns {NullishString}
+  */
   private _getColorSchemeFromLocalStorage(): NullishString {
     return localStorage.getItem('color_scheme')
   }
 
+  /**
+   * Private method: [_saveColorSchemeToLocalStorage()]
+   * @param {string} colorScheme
+   * @returns {void}
+  */
   private _saveColorSchemeToLocalStorage(colorScheme: string): void {
     localStorage.setItem('color_scheme', colorScheme)
   }
 
-  private _setActiveControl({ controls, colorScheme }: ISetActiveControlProps): void {
+  /**
+   * Private method: [_setActiveControl()]
+   * @param {HTMLElement[]} ISetActiveControlProps.controls
+   * @param {string} ISetActiveControlProps.colorScheme
+   * @returns {void}
+  */
+  private _setActiveControl({ controls, colorScheme }: _.ISetActiveControlProps): void {
     const currActiveControl: UndefinedishHTMLElem = controls.find((control: HTMLElement) => control.classList.contains('is-active'))
     const nextActiveControl: UndefinedishHTMLElem = controls.find((control: HTMLElement) => control.getAttribute('data-color-scheme') === colorScheme)
 
@@ -44,6 +67,11 @@ export default class ColorSchemeSwitcher {
     nextActiveControl?.classList.add('is-active')
   }
 
+  /**
+   * Private method: [_setColorScheme()]
+   * @param {string} colorScheme
+   * @returns {void}
+  */
   private _setColorScheme(colorScheme: string): void {
     const controls: HTMLElement[] = Array.from(this._controls)
 
@@ -54,6 +82,11 @@ export default class ColorSchemeSwitcher {
     this._updateOtherSwitchers(colorScheme)
   }
 
+  /**
+   * Private method: [_updateOtherSwitchers()]
+   * @param {string} colorScheme
+   * @returns {void}
+  */
   private _updateOtherSwitchers(colorScheme: string): void {
     const switchers: NodeListOf<HTMLElement> = document.querySelectorAll('[data-color-scheme-switcher="parent"]')
 
@@ -63,6 +96,13 @@ export default class ColorSchemeSwitcher {
     })
   }
 
+  /**
+   * @callback
+   * Private method: [_onClickControl]
+   *
+   * @param {MouseEvent} e
+   * @returns {void}
+  */
   private _onClickControl = (e: MouseEvent): void => {
     e.preventDefault()
